feat: 🎸 Member 엔티티 및 Repository 추가

+++ back/myini-domain/src/main/java/com/ssafy/myini/member/domain/MemberRepository.java
@@ -1,10 +1,9 @@
 package com.ssafy.myini.member.domain;
 
-import com.ssafy.myini.member.domain.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface MemberRepository extends JpaRepository<Member, Integer> {
+public interface MemberRepository extends JpaRepository<Member, Long> {
     Optional<Member> findByMemberName(String memberName);
 }
